image/nation/month/contents UPDATEmethod() update(pathvar idx, return VO) & token null exception

+++ BACKEND/SpringbootBackend/SpringbootYoungBlood/src/main/java/com/yb/rest/controller/ManageController.java
@@ -1,8 +1,6 @@
 package com.yb.rest.controller;
 
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 
@@ -62,7 +60,8 @@ public class ManageController {
 			@RequestParam("page") String page) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
-		ArrayList<NationDTO> list = null;
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
+		List<NationDTO> list = null;
 		Member member = new Member();
 		int pageIdx = 0;
 		if (page == null)
@@ -114,7 +113,8 @@ public class ManageController {
 		}
 		return res;
 	}
-
+	
+	/** 사용자 상품 보기 서비스 */
 	@GetMapping("/man/nation/{idx}")
 	@ApiOperation(value = "광고주의 등록 상품정보 보기 서비스.")
 	public ResponseEntity<Map<String, Object>> nationInfo(@RequestHeader(value = "Authorization") String token,
@@ -123,63 +123,55 @@ public class ManageController {
 		Map<String, Object> msg = new HashMap<String, Object>();
 		Nation nat = new Nation();
 		Monthtb mon = new Monthtb();
-		ArrayList<Route> contentsList = null;
-		ArrayList<Image> imageList = null;
+		List<Route> contentsList = null;
+		List<Image> imageList = null;
 		Member member = new Member();
 		try {
 			Claims de = MemberController.verification(token);
-//			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			member = ser.selectMemberInfo(customer);
+			
 			Owner owner = new Owner(member.getUsername(), member.getCompany(), member.getGrade());
 			int grade = ser.searchGrade(customer);
-			// 유저객체추가.
 			if (grade == 1) {
-				// 관리자는 다 볼 수 있고...
 				nat = ser.nationInfo(idx);
 				mon = ser.monthInfo(idx);
 				contentsList = ser.contentsInfo(idx);
 				imageList = ser.imagesInfo(idx);
-				msg.put("resmsg", "조회성공");
 				msg.put("nation", nat);
 				msg.put("month", mon);
 				msg.put("contents", contentsList);
 				msg.put("images", imageList);
 				msg.put("owner", owner);
+				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else if (grade >= 2) {
 				Nation natCus = ser.selectNationCustomer(customer);
 				int cusIdx = Integer.parseInt(natCus.getCustomer());
 				if (cusIdx == customer) {
 					natCus = ser.nationInfo(idx);
-					// monthtb에 nation=#{idx}로 selectOne
 					nat = ser.nationInfo(idx);
 					mon = ser.monthInfo(idx);
-
-					// img랑 contents는 selectList로 받아와야하고.
 					contentsList = ser.contentsInfo(idx);
 					imageList = ser.imagesInfo(idx);
-					
-					//counsel List
 					List<Counsel> counlist = ser.selectCounsellist(idx);
 
-					msg.put("resmsg", "조회성공");
 					msg.put("nation", nat);
 					msg.put("month", mon);
 					msg.put("contents", contentsList);
 					msg.put("images", imageList);
 					msg.put("owner", owner);
 					msg.put("counselList", counlist);
+					res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 				} else {
-					msg.put("resmsg", "권한없음");
+					res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
 				}
 			}
-			res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 		} catch (Exception e) {
-			msg.put("resmsg", "조회실패");
+			msg.put("resmsg", e.getMessage());
+			System.out.println(e.getMessage());
 			res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.NOT_FOUND);
 		}
-
 		return res;
 	}
 
@@ -190,15 +182,15 @@ public class ManageController {
 			@RequestBody Nation nat) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade > 0) {
-				boolean resProduct = ser.nationinsert(nat.getEn_name(), nat.getKo_name(), nat.getDust(),
+				ser.nationinsert(nat.getEn_name(), nat.getKo_name(), nat.getDust(),
 						nat.getContinents(), customer + "", nat.getWeight(), nat.getSpeech(), nat.getPrice(),
 						nat.getS_date(), nat.getF_date());
 				int last = Integer.MIN_VALUE;
@@ -216,7 +208,6 @@ public class ManageController {
 			System.out.println(e.getMessage());
 			res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
 		}
-
 		return res;
 	}
 
@@ -227,16 +218,15 @@ public class ManageController {
 			@PathVariable("idx") String idx) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
-			System.out.println(customer);
+
 			if (grade == 1) {
-				boolean resDelete = ser.nationdelete(idx, customer + "");
-				msg.put("resvalue", resDelete);
+				ser.nationdelete(idx, customer + "");
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -255,9 +245,9 @@ public class ManageController {
 	public ResponseEntity<Map<String, Object>> nationUpdate(@RequestHeader(value = "Authorization") String token, @RequestBody Nation nat, @PathVariable("idx") int idx) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			//msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
@@ -293,9 +283,8 @@ public class ManageController {
 	@ApiOperation(value = "상품 월별정보(month) 리스트")
 	public ResponseEntity<Map<String, Object>> monthInfo(@RequestHeader(value = "Authorization") String token) {
 		ResponseEntity<Map<String, Object>> res = null;
-		Monthtb mt = new Monthtb();
 		Map<String, Object> msg = new HashMap<String, Object>();
-		ArrayList<Monthtb> list = null;
+		List<Monthtb> list = null;
 		try {
 			Claims de = MemberController.verification(token);
 			msg.put("username", de.get("username"));
@@ -328,15 +317,15 @@ public class ManageController {
 			@RequestBody Monthtb montb) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade > 0) {
-				boolean resMonth = ser.insertMonthtb(montb);
+				ser.insertMonthtb(montb);
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -349,21 +338,22 @@ public class ManageController {
 	}
 
 	/** 상품 월별정보 수정 */
-	@PutMapping("/man/monthtb/update")
+	@PutMapping("/man/monthtb/update/{idx}")
 	@ApiOperation(value = "상품 월별정보(month) 수정")
-	public ResponseEntity<Map<String, Object>> monthUpdate(@RequestHeader(value = "Authorization") String token, @RequestBody Monthtb montb) {
+	public ResponseEntity<Map<String, Object>> monthUpdate(@RequestHeader(value = "Authorization") String token, @RequestBody Monthtb montb, @PathVariable("idx") int idx) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade == 1) {
-				boolean resMonth = ser.updateMonthtb(montb);
-				msg.put("resvalue", resMonth);
+				montb.setIdx(idx);
+				ser.updateMonthtb(montb);
+				msg.put("updateMonth", montb);
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -383,15 +373,15 @@ public class ManageController {
 			@PathVariable("nation") String nation) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade == 1) {
-				boolean resDelete = ser.deleteMonthtb(Integer.parseInt(nation));
+				ser.deleteMonthtb(Integer.parseInt(nation));
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -410,10 +400,9 @@ public class ManageController {
 	public ResponseEntity<Map<String, Object>> ContentsList(@RequestHeader(value = "Authorization") String token) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
-
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
@@ -445,9 +434,9 @@ public class ManageController {
 			@RequestBody List<Route> route) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
@@ -468,21 +457,23 @@ public class ManageController {
 	}
 
 	/** 콘텐츠 정보 수정하기 */
-	@PutMapping("/man/contents/update")
+	@PutMapping("/man/contents/update/{idx}")
 	@ApiOperation(value = "상품 콘텐츠 정보(contents) 수정")
 	public ResponseEntity<Map<String, Object>> ContentsUpdate(@RequestHeader(value = "Authorization") String token,
-			@RequestBody Route route) {
+			@RequestBody Route route, @PathVariable("idx") int idx) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade == 1) {
+				route.setIdx(idx+"");
 				adser.updateRoutes(route);
+				msg.put("inputContents", route);
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -502,6 +493,7 @@ public class ManageController {
 			@PathVariable("idx") int idx, @PathVariable("nation") int nation) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
 			msg.put("username", de.get("username"));
@@ -532,10 +524,10 @@ public class ManageController {
 	public ResponseEntity<Map<String, Object>> imageList(@RequestHeader(value = "Authorization") String token) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
-		ArrayList<Image> list = null;
+		List<Image> list = null;
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
@@ -564,6 +556,7 @@ public class ManageController {
 			@RequestBody List<Image> imgs) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
 			msg.put("username", de.get("username"));
@@ -590,21 +583,23 @@ public class ManageController {
 	}
 
 	/** 이미지 정보 수정 */
-	@PutMapping("/man/image/update")
+	@PutMapping("/man/image/update/{idx}")
 	@ApiOperation(value = "이미지 정보(image) 수정")
 	public ResponseEntity<Map<String, Object>> imageUpdate(@RequestHeader(value = "Authorization") String token,
-			@RequestBody Image img) {
+			@RequestBody Image img, @PathVariable("idx") int idx) {
 		ResponseEntity<Map<String, Object>> res = null;
 		Map<String, Object> msg = new HashMap<String, Object>();
+		if(token=="" || token==null) return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED); 
 		try {
 			Claims de = MemberController.verification(token);
-			msg.put("username", de.get("username"));
 			String username = (String) de.get("username");
 			int customer = ser.getIdx(username);
 			int grade = ser.searchGrade(customer);
 
 			if (grade == 1) {
-				boolean resImg = ser.updateImagetb(img);
+				img.setIdx(idx+"");
+				ser.updateImagetb(img);
+				msg.put("update", img);
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				return new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
@@ -632,7 +627,7 @@ public class ManageController {
 			int grade = ser.searchGrade(customer);
 
 			if (grade == 1) {
-				boolean resDelete = ser.deleteImagetb(Integer.parseInt(idx));
+				ser.deleteImagetb(Integer.parseInt(idx));
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.OK);
 			} else {
 				res = new ResponseEntity<Map<String, Object>>(msg, HttpStatus.UNAUTHORIZED);
